Pull out cudnn test into its own file.
