+++ tests/test_cudnn.h
@@ -1,0 +1,8 @@
+#ifndef TEST_CUDNN_H
+#define TEST_CUDNN_H
+
+#include "tests/min_unit.h"
+
+MIN_UNIT_TEST_FUNC(test_matmul_small_cudnn);
+
+#endif /* TEST_CUDNN_H */
